Move functions into BugsnagConfig to avoid lint failures
