+++ features/fixtures/mazerunner/jvm-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/ReportCacheScenario.kt
@@ -2,6 +2,7 @@ package com.bugsnag.android.mazerunner.scenarios
 
 import android.content.Context
 import com.bugsnag.android.Configuration
+import com.bugsnag.android.mazerunner.disableAllDelivery
 
 /**
  * Sends an unhandled exception which is cached on disk to Bugsnag, then sent on a separate launch.
