+++ features/fixtures/mazerunner/jvm-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/StartupCrashFlushScenario.kt
@@ -3,6 +3,7 @@ package com.bugsnag.android.mazerunner.scenarios
 import android.content.Context
 import android.os.Handler
 import com.bugsnag.android.Configuration
+import com.bugsnag.android.mazerunner.disableAllDelivery
 
 /**
  * Generates an uncaught exception, catches it, and persists it to disc, preventing any delivery.
